refactor(mesh): split file transfer and bridge links out of MeshService
core/mesh/service.ts was 1,863 lines — one class carrying the device
registry and command channel, the chunked/streamed file transfer paths
and both self-update flows, and node provisioning + companion pairing.
The file already marked those seams with section comments; this moves
the code across them, verbatim apart from the host indirection.

- device-files.ts — DeviceFiles: acceptFileUpload/openFileDownload,
  readFileBytes and the read/write/pull/push tools, updateDeviceApp,
  updateNodeBinary, the chunked pullBytes/pushBytes and the streamed
  pullViaStream. It reaches device resolution and command dispatch
  through a DeviceFilesHost the service constructs from its own private
  methods, so nothing about the service's API widened.
- bridge-links.ts — BridgeLinks: the provisioning and pairing grant
  stores, the bridge's self-reported identity, getNodeBinary,
  makeNodeInstallLink, makeCompanionPairLink, openCompanionPair,
  bridgeReachability, openNodeInstall/Binary, and the base-URL
  derivation. MeshBridgeInfo lives here and is re-exported.
- common.ts — MeshToolResult, FS_COMMAND_TIMEOUT_MS, formatBytes, age,
  requirePath: the helpers all three need.

MeshService keeps the registry, transport fan-out, locate/wait, the
command channel, device resolution, and the model-facing summaries, plus
one-line delegations for the moved methods so every caller and test is
untouched. service.ts is 1,033 lines; the largest piece of the old file
now stands on its own at 707.
